refactor(explorer): extract useFileBrowserTree composable + fix viewer height
- Extract useFileBrowserTree(entries, opts) composable to packages/ui
  containing: treeStructure, fileCount, collapsedFolders, toggleFolder,
  formatSize — all the logic duplicated between FileBrowserTree.vue and
  SkillAssetsTree.vue

- FileBrowserTree.vue now delegates to the composable (removes ~50 lines)

- SkillAssetsTree.vue (skill editor) now delegates to the same composable
  via toRef(props, 'assets'), removing all duplicate computed/watch/ref
  logic. Skills-specific state (showNewFileInput, newFileName, etc.)
  stays local.

- Export useFileBrowserTree + UseFileBrowserTreeOptions from
  @tracepilot/ui packages (composables/index.ts + index.ts)

- Fix FileContentViewer.vue height: .fcv used height:100% which fails
  when parent height is flex-derived; changed to flex:1; min-height:0
  so the viewer fills the full right pane in the explorer
