+++ packages/ui/src/components/FileContentViewer.vue
@@ -257,7 +257,8 @@ function copyTableAsJson() {
 .fcv {
   display: flex;
   flex-direction: column;
-  height: 100%;
+  flex: 1;
+  min-height: 0;
   overflow: hidden;
 }
